+++ src/pydvl/value/least_core/naive.py
@@ -3,23 +3,21 @@ import warnings
 from typing import Optional
 
 import numpy as np
-from numpy.typing import NDArray
 
 from pydvl.utils import Utility, maybe_progress, powerset
-from pydvl.utils.status import Status
-from pydvl.value.least_core._common import (
-    _solve_egalitarian_least_core_quadratic_program,
-    _solve_least_core_linear_program,
-)
+from pydvl.value.least_core.common import LeastCoreProblem, lc_solve_problem
 from pydvl.value.results import ValuationResult
 
-__all__ = ["exact_least_core"]
+__all__ = ["exact_least_core", "lc_prepare_problem"]
 
 logger = logging.getLogger(__name__)
 
 
 def exact_least_core(
-    u: Utility, *, options: Optional[dict] = None, progress: bool = True, **kwargs
+    u: Utility,
+    *,
+    options: Optional[dict] = None,
+    progress: bool = True,
 ) -> ValuationResult:
     r"""Computes the exact Least Core values.
 
@@ -51,18 +49,28 @@ def exact_least_core(
     :return: Object with the data values and the least core value.
     """
     n = len(u.data)
-
-    # Arbitrary choice, will depend on time required, caching, etc.
-    if n > 20:
+    if n > 20:  # Arbitrary choice, will depend on time required, caching, etc.
         warnings.warn(f"Large dataset! Computation requires 2^{n} calls to model.fit()")
 
-    if options is None:
-        options = {}
+    problem = lc_prepare_problem(u, progress=progress)
+    return lc_solve_problem(
+        problem=problem, u=u, algorithm="exact_least_core", **(options or {})
+    )
 
-    powerset_size = 2**n
+
+def lc_prepare_problem(u: Utility, progress: bool = False) -> LeastCoreProblem:
+    """Prepares a linear problem with all subsets of the data
+    Use this to separate the problem preparation from the solving with
+    :func:`~pydvl.value.least_core.common.lc_solve_problem`. Useful for
+    parallel execution of multiple experiments.
+
+    See :func:`~pydvl.value.least_core.naive.exact_least_core` for argument
+    descriptions.
+    """
+    n = len(u.data)
 
     logger.debug("Building vectors and matrices for linear programming problem")
-    A_eq = np.ones((1, n))
+    powerset_size = 2**n
     A_lb = np.zeros((powerset_size, n))
 
     logger.debug("Iterating over all subsets")
@@ -80,54 +88,4 @@ def exact_least_core(
         A_lb[i, indices] = 1
         utility_values[i] = u(subset)
 
-    b_lb = utility_values
-    b_eq = utility_values[-1:]
-
-    _, subsidy = _solve_least_core_linear_program(
-        A_eq=A_eq, b_eq=b_eq, A_lb=A_lb, b_lb=b_lb, **options
-    )
-
-    values: Optional[NDArray[np.float_]]
-
-    if subsidy is None:
-        logger.debug("No values were found")
-        status = Status.Failed
-        values = np.empty(n)
-        values[:] = np.nan
-        subsidy = np.nan
-
-        return ValuationResult(
-            algorithm="exact_least_core",
-            status=status,
-            values=values,
-            subsidy=subsidy,
-            stderr=None,
-            data_names=u.data.data_names,
-        )
-
-    values = _solve_egalitarian_least_core_quadratic_program(
-        subsidy,
-        A_eq=A_eq,
-        b_eq=b_eq,
-        A_lb=A_lb,
-        b_lb=b_lb,
-        **options,
-    )
-
-    if values is None:
-        logger.debug("No values were found")
-        status = Status.Failed
-        values = np.empty(n)
-        values[:] = np.nan
-        subsidy = np.nan
-    else:
-        status = Status.Converged
-
-    return ValuationResult(
-        algorithm="exact_least_core",
-        status=status,
-        values=values,
-        subsidy=subsidy,
-        stderr=None,
-        data_names=u.data.data_names,
-    )
+    return LeastCoreProblem(utility_values, A_lb)